+++ test/DelimitedDataParser.Test/Properties/AssemblyInfo.cs
@@ -5,5 +5,4 @@ using System.Runtime.InteropServices;
 [assembly: AssemblyTitle("Delimited Data Parser Tests")]
 [assembly: AssemblyProduct("Delimited Data Parser")]
 [assembly: ComVisible(false)]
-[assembly: CLSCompliant(true)]
 [assembly: AssemblyVersion("3.2.*")]